move DataSourceSelectionSection to actions/configuration with the other ones
