+++ front/components/assistant_builder/actions/ProcessAction.tsx
@@ -16,9 +16,9 @@ import {
 } from "@dust-tt/sparkle";
 import React, { useEffect, useState } from "react";
 
+import DataSourceSelectionSection from "@app/components/assistant_builder/actions/configuration/DataSourceSelectionSection";
 import { TimeUnitDropdown } from "@app/components/assistant_builder/actions/TimeDropdown";
 import AssistantBuilderDataSourceModal from "@app/components/assistant_builder/AssistantBuilderDataSourceModal";
-import DataSourceSelectionSection from "@app/components/assistant_builder/DataSourceSelectionSection";
 import type {
   AssistantBuilderActionConfiguration,
   AssistantBuilderProcessConfiguration,
